+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/ScenarioTestingApiHttpService.scala
@@ -8,18 +8,17 @@ import pl.touk.nussknacker.engine.definition.test.TestInfoProvider.{
   ScenarioTestDataGenerationError,
   TestingCapabilitiesError
 }
-import pl.touk.nussknacker.restmodel.BaseEndpointDefinitions
 import pl.touk.nussknacker.restmodel.validation.PrettyValidationErrors
 import pl.touk.nussknacker.restmodel.validation.ValidationResults.ValidationErrors
 import pl.touk.nussknacker.security.Permission
 import pl.touk.nussknacker.security.Permission.Permission
 import pl.touk.nussknacker.ui.api.BaseHttpService.CustomAuthorizationError
-import pl.touk.nussknacker.ui.api.ScenarioTestingApiHttpService.TestingError
-import pl.touk.nussknacker.ui.api.ScenarioTestingApiHttpService.TestingError._
-import pl.touk.nussknacker.ui.api.ScenarioTestingApiHttpService.TestingError.BadRequestTestingError._
-import pl.touk.nussknacker.ui.api.ScenarioTestingApiHttpService.TestingError.NotFoundTestingError._
 import pl.touk.nussknacker.ui.api.description.NodesApiEndpoints.Dtos.ParametersValidationResultDto
-import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.{ResultsWithCountsDto, ScenarioTestData}
+import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.{
+  ResultsWithCountsDto,
+  ScenarioTestData,
+  TestingError
+}
 import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.Capabilities.{
   CapabilityStatus,
   NotAvailableReason,
@@ -28,9 +27,11 @@ import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.Capabilities.
 }
 import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.Capabilities.TestCapabilityDetails.TestWithParametersDetails
 import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.Test.{SkipResultsPerNode, SkipResultsPerTransition}
+import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.TestingError._
+import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.TestingError.BadRequestTestingError._
+import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.TestingError.NotFoundTestingError._
 import pl.touk.nussknacker.ui.api.description.scenarioTesting.ScenarioTestingApiEndpoints
 import pl.touk.nussknacker.ui.api.utils.ScenarioHttpServiceExtensions
-import pl.touk.nussknacker.ui.api.utils.ValidationErrorOps.ValidationErrorOps
 import pl.touk.nussknacker.ui.process.ProcessService
 import pl.touk.nussknacker.ui.process.processingtype.provider.ProcessingTypeDataProvider
 import pl.touk.nussknacker.ui.process.test.PreliminaryScenarioTestDataSerDe.SerializationError
@@ -38,9 +39,6 @@ import pl.touk.nussknacker.ui.process.test.ScenarioTestService
 import pl.touk.nussknacker.ui.process.test.ScenarioTestService.GenerateTestDataError
 import pl.touk.nussknacker.ui.security.api.{AuthManager, LoggedUser}
 import pl.touk.nussknacker.ui.validation.ParametersValidator
-import sttp.model.StatusCode.NotFound
-import sttp.tapir.{oneOfVariant, plainBody, Codec, CodecFormat, EndpointOutput}
-import sttp.tapir.EndpointIO.Example
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -280,89 +278,5 @@ class ScenarioTestingApiHttpService(
           case false => Left(noPermissionError)
         }
     )
-
-}
-
-object ScenarioTestingApiHttpService {
-
-  sealed trait TestingError
-
-  object TestingError {
-
-    final case object NoPermission extends TestingError with CustomAuthorizationError
-
-    sealed trait BadRequestTestingError extends TestingError
-
-    object BadRequestTestingError {
-      final case class TooManyCharactersGenerated(length: Int, limit: Int)    extends BadRequestTestingError
-      final case class TooManySamplesRequested(maxSamples: Int)               extends BadRequestTestingError
-      final case class ScenarioGraphValidationError(errors: ValidationErrors) extends BadRequestTestingError
-
-      implicit val badRequestTestingErrorCodec: Codec[String, BadRequestTestingError, CodecFormat.TextPlain] = {
-        BaseEndpointDefinitions.toTextPlainCodecSerializationOnly[BadRequestTestingError] {
-          case ScenarioGraphValidationError(errors) =>
-            errors.toHumanReadableMessage
-          case TooManyCharactersGenerated(length, limit) =>
-            TestingApiErrorMessages.generatedTestData.tooManyCharacters(length, limit)
-          case TooManySamplesRequested(maxSamples) =>
-            TestingApiErrorMessages.generatedTestData.requestedTooManySamplesToGenerate(maxSamples)
-        }
-      }
-
-    }
-
-    sealed trait NotFoundTestingError extends TestingError
-
-    object NotFoundTestingError {
-      final case class NoScenario(scenarioName: ProcessName) extends NotFoundTestingError
-      final case object NoDataGenerated                      extends NotFoundTestingError
-      final case object NoSourcesWithTestDataGeneration      extends NotFoundTestingError
-
-      implicit val notFoundTestingErrorCodec: Codec[String, NotFoundTestingError, CodecFormat.TextPlain] = {
-        BaseEndpointDefinitions.toTextPlainCodecSerializationOnly[NotFoundTestingError] {
-          case NoScenario(scenarioName) => s"No scenario ${scenarioName.value} found"
-          case NoDataGenerated          => TestingApiErrorMessages.generatedTestData.couldNotProvideTestDataSample
-          case NoSourcesWithTestDataGeneration =>
-            TestingApiErrorMessages.generatedTestData.noSourcesWithTestDataGeneration
-        }
-      }
-
-      implicit val noScenarioCodec: Codec[String, NoScenario, CodecFormat.TextPlain] = {
-        BaseEndpointDefinitions.toTextPlainCodecSerializationOnly[NoScenario](e =>
-          s"No scenario ${e.scenarioName} found"
-        )
-      }
-
-    }
-
-    final case class UnsupportedOperation(message: String) extends TestingError
-
-    implicit val UnsupportedOperationCodec: Codec[String, UnsupportedOperation, CodecFormat.TextPlain] = {
-      BaseEndpointDefinitions.toTextPlainCodecSerializationOnly[UnsupportedOperation](_.message)
-    }
-
-    final case class ErrorResult(message: String) extends TestingError
-
-    implicit val ErrorResultCodec: Codec[String, ErrorResult, CodecFormat.TextPlain] = {
-      BaseEndpointDefinitions.toTextPlainCodecSerializationOnly[ErrorResult](_.message)
-    }
-
-  }
-
-  object Examples {
-
-    val noScenarioExample: Example[NoScenario] = Example.of(
-      summary = Some("No scenario {scenarioName} found"),
-      value = NoScenario(ProcessName("'example scenario'"))
-    )
-
-    // TODO
-    val noScenarioErrorOutput: EndpointOutput.OneOfVariant[NoScenario] =
-      oneOfVariant(
-        NotFound,
-        plainBody[NoScenario].example(noScenarioExample)
-      )
-
-  }
 
 }
